+++ src/main/kotlin/BitbucketWatcherApplication.kt
@@ -1,0 +1,14 @@
+package bitbucketwatcher
+
+import org.springframework.boot.autoconfigure.SpringBootApplication
+import org.springframework.boot.runApplication
+import org.springframework.scheduling.annotation.EnableScheduling
+
+@SpringBootApplication
+@EnableScheduling
+class BitbucketWatcherApplication
+
+fun main(args: Array<String>) {
+    @Suppress("SpreadOperator") // no performance issue here
+    runApplication<BitbucketWatcherApplication>(*args)
+}
