+++ src/main/kotlin/client/TeamsClient.kt
@@ -1,0 +1,14 @@
+package client
+
+import feign.Headers
+import feign.RequestLine
+import model.TeamsMessage
+import org.springframework.stereotype.Component
+
+@Component
+interface TeamsClient {
+
+    @Headers("Content-Type: application/json; charset=utf-8")
+    @RequestLine("POST /")
+    fun postMessage(teamsMessage: TeamsMessage)
+}
