--- src/main/kotlin/de/otto/bitbucketwatcher/BitbucketWatcherApplication.kt
@@ -1,14 +1,0 @@
-package de.otto.bitbucketwatcher
-
-import org.springframework.boot.autoconfigure.SpringBootApplication
-import org.springframework.boot.runApplication
-import org.springframework.scheduling.annotation.EnableScheduling
-
-@SpringBootApplication
-@EnableScheduling
-class BitbucketWatcherApplication
-
-fun main(args: Array<String>) {
-    @Suppress("SpreadOperator") // no performance issue here
-    runApplication<BitbucketWatcherApplication>(*args)
-}
