--- src/main/kotlin/de/otto/bitbucketwatcher/logger/LogUtils.kt
@@ -1,5 +1,0 @@
-package de.otto.bitbucketwatcher.logger
-
-import org.slf4j.LoggerFactory
-
-fun getLogger(classInstance: Class<out Any>) = LoggerFactory.getLogger(classInstance) ?: error("Could not get Logger!")
