+++ src/main/kotlin/logger/LogUtils.kt
@@ -1,0 +1,5 @@
+package logger
+
+import org.slf4j.LoggerFactory
+
+fun getLogger(classInstance: Class<out Any>) = LoggerFactory.getLogger(classInstance) ?: error("Could not get Logger!")
